move mock data and more storybook deletions
