+++ locale/es.json
@@ -688,7 +688,6 @@
 	"pages.rooms.importCourse.btn.confirm": "Importar",
 	"pages.rooms.importCourse.error": "El código del curso no está en uso.",
 	"pages.termsofuse.title": "Condiciones de uso y política de privacidad",
-	"storybook.lang": "¡Estoy en inglés ahora!",
 	"utils.adminFilter.class.title": "Clase(s)",
 	"utils.adminFilter.consent": "Formulario de consentimiento:",
 	"utils.adminFilter.consent.label.missing": "Usuario creado",
